pyproject.toml
[tool.black]
line-length = 120
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
| sparklehub/openapi_client
)
'''

[tool.poetry]
name = "sparklehub"
version = "0.1.0"
description = ""
authors = ["Alexander Shishenko <alex@shishenko.com>"]
packages = [
    { include = "sparklehub" }
]
scripts = { sparklehub = 'sparklehub.main:main' }

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.1"
cryptography = "^2.9.2"
httpx = "^0.12.1"
tqdm = "^4.46.0"
pure25519 = "^0.0.1"
pyobjc-framework-Security = {version = "^6.2", platform = "darwin"}
colorama = "^0.4.3"
urllib3 = "^1.15"
six = "^1.10"
python-dateutil = "^2.8.1"
certifi = ">=2020.4.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
flake8 = "^3.7.9"
ipython = "^7.13.0"
pre-commit = "^2.1"
pytest = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
